+++ src/helpers/Table.php
@@ -18,6 +18,7 @@ abstract class Table extends CraftTable
     public const FORMIE_FORM_TEMPLATES = '{{%formie_formtemplates}}';
     public const FORMIE_FORM_GROUPS = '{{%formie_formgroups}}';
     public const FORMIE_INTEGRATIONS = '{{%formie_integrations}}';
+    public const FORMIE_CAPTCHA_PROVIDERS = '{{%formie_captcha_providers}}';
     public const FORMIE_NOTIFICATIONS = '{{%formie_notifications}}';
     public const FORMIE_PAYMENTS = '{{%formie_payments}}';
     public const FORMIE_PAYMENT_PLANS = '{{%formie_payments_plans}}';