+++ src/records/CaptchaProvider.php
@@ -1,0 +1,17 @@
+<?php
+namespace verbb\formie\records;
+
+use verbb\formie\helpers\Table;
+
+use craft\db\ActiveRecord;
+
+class CaptchaProvider extends ActiveRecord
+{
+    // Static Methods
+    // =========================================================================
+
+    public static function tableName(): string
+    {
+        return Table::FORMIE_CAPTCHA_PROVIDERS;
+    }
+}